Enable use of tiddlywebplugins.virtualhosting
If you come in on a virtualhost that is not the same as the
main server_host, then attempt to load up the recipes associated
with that host. Thenceforth URLs generated on HTML and tiddlywiki
representations will use the correct urls, enabling saving.

This does nothing related to authentication and cookies.