Add support for creating a user via a POST to /users.
The handler listening on /users is called post_user. It accepts
a JSON dictionary with 'username' and 'password' keys, creating a user.

If something other than JSON is senting, a HTTP 415 is returned.
If the user is not a role ADMIN, 403.
If data is missing from the JSON, 400.
If the user already exists, 409.

If the user is created, 201 is returned.