Stub in a tiddlywiki based challenger.
The idea is that if the system wants to raise a challenger (say
you go to a private recipe) this challenger will direct the browser
to a tiddlywiki that is created from a recipe named '_challenger',
which contains the bag '_challenger'.

It is expect that at some point other bags may be added to the
recipe (via instance.py) and that tiddlers containing the necessary
macros, plugins and content will be added to the bags to present
a useful interface. Those tiddlers are _not_ a part of this commit.

Those macros, for the time being, will be expected to do a POST
to the simple_cookie challenger, of which this challenger is a subclass.
The structure and URL of that challenger can be seen by looking at the
form at:

  http://tiddlyweb.peermore.com/wiki/challenge/cookie_form

Adjustments will be required to make sure the action attribute of
the form used by the macro has the right URL.

This commit updates the instance_config so that it will use
only 'tiddlywebplugins.tiddlyspace.challenger' as its auth_system.