Stub in the /dashboard  handler.
The /dashboard handler loads up and presents the
_dashboard recipe, which is made up of the
_dashboard bag, whose content comes from
src/dashboard/split.recipe, all of which is
referenced in instance.py.

Note the XXX comment in handler.py about authentication.
Right now we can't do auth because the challenger
sends the _challenger wiki, but it does not have the
login macros. When auth is enabled, the line following
the XXX comment must be removed.